refactor(mix_generator): share widget call rendering
